Find max diameter of a full binary tree
In a binary tree, diameter sum between two leaf nodes is defined as sum of all the nodes in the unique path when travelling from one leaf to the other. Assume that the tree is a complete binary tree and every leaf node is at the same depth from root of the tree. Find the value of maximum diameter sum in a binary tree. Note that the maximum diameter may be a single leaf node as well (since a single leaf node is also a valid diameter the trivial path of length 0 from the leaf node to itself).